@@ -1,15 +1,73 @@
 [project]
-name = "kimi-vendor-verfier"
+name = "kimi-vendor-verifier"
 version = "0.1.0"
-description = "Add your description here"
+description = "Tool Calls Validator for Kimi K2 model evaluation"
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.14"
+authors = [
+    {name = "Moonshot AI", email = "shijuanfeng@moonshot.cn"},
+]
+license = {text = "MIT"}
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.14",
+]
 dependencies = [
     "httpx>=0.28.1",
+    "huggingface-hub>=0.36.0",
     "jsonschema>=4.25.1",
     "loguru>=0.7.3",
     "megfile>=4.2.4",
     "openai>=1.108.1",
     "tqdm>=4.67.1",
     "transformers>=4.57.1",
+]
+
+[project.optional-dependencies]
+dev = [
+    "ruff>=0.8.0",
+]
+
+[project.scripts]
+kimi-vendor-verifier = "kimi_vendor_verifier.cli:main"
+k2vv = "kimi_vendor_verifier.cli:main"
+
+[project.urls]
+Homepage = "https://github.com/MoonshotAI/K2-Vendor-Verifier"
+Repository = "https://github.com/MoonshotAI/K2-Vendor-Verifier"
+Issues = "https://github.com/MoonshotAI/K2-Vendor-Verifier/issues"
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.ruff]
+target-version = "py314"
+
+# enable running under Python 3.14 without "this may be too new" warnings
+preview = true
+
+[tool.ruff.lint]
+select = [
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "B",  # flake8-bugbear
+    "C4", # flake8-comprehensions
+    "UP", # pyupgrade
+]
+ignore = [
+    "E501",  # line too long
+#    "B008",  # do not perform function calls in argument defaults
+#    "C901",  # too complex
 ]
